Moved DNSSeedData and Consensus to their own classes. Consolidated some of the base58 creation methods in Network
